PD-597: Update structure of mongo-select
