+++ packages/mongo-nav/src/org-nav/OrgNav.tsx
@@ -5,7 +5,7 @@ import Badge, { Variant } from '@leafygreen-ui/badge';
 import IconButton from '@leafygreen-ui/icon-button';
 import Icon from '@leafygreen-ui/icon';
 import UserMenu from '../user-menu';
-import { css, cx } from '@leafygreen-ui/emotion';
+import { css } from '@leafygreen-ui/emotion';
 import { uiColors } from '@leafygreen-ui/palette';
 import { LogoMark } from '@leafygreen-ui/logo';
 import { useViewportSize } from '@leafygreen-ui/hooks';
@@ -40,16 +40,6 @@ const navContainer = css`
   background-color: white;
   border-bottom: 1px solid ${uiColors.gray.light2};
   color: ${uiColors.gray.dark3};
-`;
-
-const orgSelectContainer = css`
-  margin-left: 20px;
-  margin-right: 20px;
-`;
-
-const disabledOrgSelect = css`
-  cursor: default;
-  pointer-events: none;
 `;
 
 const supportContainer = css`
@@ -253,11 +243,10 @@ function OrgNav({
       </Tooltip>
 
       <OrgSelect
-        className={cx(orgSelectContainer, { [disabledOrgSelect]: disabled })}
         data={data}
         current={current}
         constructOrganizationURL={constructOrganizationURL}
-        urls={urls}
+        urls={urls.mongoSelect}
         onChange={onOrganizationChange}
         isActive={activeNav === ActiveNavElement.OrgNavOrgSettings}
         loading={!current}
